+++ index.js
@@ -1,18 +1,19 @@
 /** @typedef {import('./src/index.vue').ButtonInfo} ButtonInfo */
 
 
-export { $title as title } from './src/index.vue';
-export { $content as content } from './src/index.vue';
-export { $cancel as cancel } from './src/index.vue';
+export { $title } from './src/index.vue';
+export { $content } from './src/index.vue';
+export { $cancel } from './src/index.vue';
 
-export { $button1 as button1 } from './src/index.vue';
-export { $button2 as button2 } from './src/index.vue';
-export { $button3 as button3 } from './src/index.vue';
+export { $button1 } from './src/index.vue';
+export { $button2 } from './src/index.vue';
+export { $button3 } from './src/index.vue';
 
-export { $colorTop as colorTop } from './src/index.vue';
+export { $colorTop } from './src/index.vue';
 
-export { $show as show } from './src/index.vue';
-export { $waiter as waiter } from './src/index.vue';
+export { $showing } from './src/index.vue';
+export { $resolver } from './src/index.vue';
+
 
 export { $alert } from './src/index.vue';
 export { $quest } from './src/index.vue';
@@ -20,5 +21,6 @@ export { $quest3 } from './src/index.vue';
 
 export { $okay } from './src/index.vue';
 export { $fail } from './src/index.vue';
+
 
 export { install } from './src/index.vue';